Remove elm-stuff from repo.
Elm-stuff leaked into the git repository before the .gitignore was updated. As
that folder is automatically generated upon compilation, it's redundant to ship
it alongside the rest of the repo. Hence, I'm pushing a delete of the entire
folder so that from now on changes to it won't be included in the git history.